Adding the schema.org codemeta-compliant json-ld
Daniel Garijo and I just completed creating a schema.org json-ld that is codemeta-compliant, which means the json-ld validates in the Schema.org validator. We started with the full schema.org json-ld already in the repository and removed terms used to replace codemeta terms (unless they have the same name in Schema.org). Other terms used in Schema.org are allowed (e.g. using codemeta:developmentStatus *and* creativeWorkStatus).
There are no critical issues for the software item in Google's RichResults tool. However, there are critical issues for the datasets and the review indicated in the record. The issues associated with the dataset can be addressed by pulling that metadata from the DataCite API. The review item is fictional, but simply adding the author for an actual review would resolve the issue. Removing the field does not cause addition issues (critical or noncritical).
Please note that although the SunPy software package is used as the example in this record, some of the metadata here is fictional (e.g. ORCiD = 1234-1234-1234-1234).