Show how license can be expanded
The license file on the code repository might be edited compared to the spdx version, so I slightly expanded the license field to be a CreativeWork instead of a URL so that both the spdx identifier and the URL to the license file can be included. I also added the name field in case a landing page wanted to have text to represent the link.